Original Request UGC details function has been deprecated.
New version supports arrays of data and dataset filtering, not using that yet but will in the future.

Also changes UGCFile_ID to be uint64 to match the new backend for steam, it used to be uint32.


Former-commit-id: 6cdb799eb0256fbad68c44d6bc597d11c10abe44

+++ AdvancedSteamSessions/Source/AdvancedSteamSessions/Classes/AdvancedSteamWorkshopLibrary.h
@@ -48,14 +48,14 @@ struct FBPSteamWorkshopID
 
 public:
 
-	uint32 SteamWorkshopID;
+	uint64 SteamWorkshopID;
 
 	FBPSteamWorkshopID()
 	{
 
 	}
 
-	FBPSteamWorkshopID(uint32 ID)
+	FBPSteamWorkshopID(uint64 ID)
 	{
 		SteamWorkshopID = ID;
 	}
